+++ packages/localnet/src/zetachainWithdraw.ts
@@ -1,12 +1,13 @@
 import { ethers, NonceManager } from "ethers";
-import { handleOnRevertZEVM } from "./handleOnRevertZEVM";
-import { log, logErr } from "./log";
+import { zetachainOnRevert } from "./zetachainOnRevert";
+import { log } from "./log";
 import { deployOpts } from "./deployOpts";
 import * as ZRC20 from "@zetachain/protocol-contracts/abi/ZRC20.sol/ZRC20.json";
 import { solanaWithdraw } from "./solanaWithdraw";
-import { FixedNumber } from "ethers";
+import { evmTSSTransfer } from "./evmTSSTransfer";
+import { evmCustodyWithdraw } from "./evmCustodyWithdraw";
 
-export const handleOnZEVMWithdrawn = async ({
+export const zetachainWithdraw = async ({
   evmContracts,
   tss,
   provider,
@@ -34,56 +35,24 @@ export const handleOnZEVMWithdrawn = async ({
     (coin: any) => coin.zrc20_contract_address === zrc20
   )?.foreign_chain_id;
 
-  const getERC20ByZRC20 = (zrc20: string) => {
-    const foreignCoin = foreignCoins.find(
-      (coin: any) => coin.zrc20_contract_address === zrc20
-    );
-    if (!foreignCoin) {
-      logErr(chainID, `Foreign coin not found for ZRC20 address: ${zrc20}`);
-      return;
-    }
-    return foreignCoin.asset;
-  };
-
   const amount = args[4];
   try {
-    const receiver = args[2];
     (tss as NonceManager).reset();
     const zrc20Contract = new ethers.Contract(zrc20, ZRC20.abi, deployer);
     const coinType = await zrc20Contract.COIN_TYPE();
     const isGasToken = coinType === 1n;
     const isERC20orZETA = coinType === 2n;
+    const isSolana = chainID === "901";
 
-    if (chainID === "901") {
-      console.log("amount", amount / BigInt(10 ** 9));
-      await solanaWithdraw(
-        ethers.toUtf8String(receiver),
-        amount / BigInt(10 ** 9)
-      );
+    if (isSolana) {
+      const receiver = ethers.toUtf8String(args[2]);
+      const amountFormatted = amount / BigInt(10 ** 9);
+      await solanaWithdraw(receiver, amountFormatted);
     } else {
       if (isGasToken) {
-        const tx = await tss.sendTransaction({
-          to: receiver,
-          value: amount,
-          ...deployOpts,
-        });
-        await tx.wait();
-        log(
-          chainID,
-          `Transferred ${ethers.formatEther(
-            amount
-          )} native gas tokens from TSS to ${receiver}`
-        );
+        await evmTSSTransfer({ tss, args, foreignCoins });
       } else if (isERC20orZETA) {
-        const erc20 = getERC20ByZRC20(zrc20);
-        const tx = await evmContracts[chainID].custody
-          .connect(tss)
-          .withdraw(receiver, erc20, amount, deployOpts);
-        await tx.wait();
-        log(
-          chainID,
-          `Transferred ${amount} ERC-20 tokens from Custody to ${receiver}`
-        );
+        await evmCustodyWithdraw({ evmContracts, tss, args, foreignCoins });
       }
     }
   } catch (err: any) {
@@ -91,7 +60,7 @@ export const handleOnZEVMWithdrawn = async ({
       throw new Error(err);
     }
     const revertOptions = args[9];
-    return await handleOnRevertZEVM({
+    return await zetachainOnRevert({
       revertOptions,
       err,
       provider,
